requirements.txt
# Web framework
fastapi==0.110.1
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
geoalchemy2==0.15.1

# Cache
redis[asyncio]==5.0.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.2.1
pydantic==2.6.4
python-dotenv==1.0.1

# HTTP client (PSP calls)
httpx==0.27.0

# Monitoring
newrelic==9.9.0

# Utilities
python-multipart==0.0.9

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0
factory-boy==3.3.0
